ansible: Default to https git URLs, public repositories, & dev branch
The intention of these defaults is to work for what I believe is the
most common case - a validators/stakers installing a node for mainnet.
This allows the minimal inventories included (e.g.
scripts/ansible/inventories/raspberrypi.yml) to be used unmodified for
installing/upgrading a node.

However this might come in conflict with use cases of the Ava Labs team.
For those I propose overriding these variables in any inventories used
by the team.