logging: Add syslog format, use it with systemd
When a program is run under systemd stdout and stderr are (by default) sent
to the systemd journal. Each line is logged at info severity.

A program can detect the presence of the journal on stdout/stderr and modify
its behaviour to log at other severities by prefixing each line with "<n>",
where n is an integer between 0 and 7.

See:
- https://www.freedesktop.org/software/systemd/man/sd-daemon.html
- https://systemd.io/JOURNAL_NATIVE_PROTOCOL/#automatic-protocol-upgrading