Pull in dependent repositories (#123)
Maintenance on multiple repos is harder than keeping everything in
one place. Now that we have workspaces, just grab the latest copy
of each dependency and check it in here.

There is some loss of history for the original repos, but it didn't
look like there was much.

Removed all rustfmt.toml files, and ran 'cargo fmt' on these before
committing.