Ensure scripts/lint.sh failure when used with incompatible grep
scripts/lint.sh requires the use of GNU grep due to dependency on the
`--perl-regexp` flag, and the default version installed on MacOS is
not compatible. Previously this compatibility problem wasn't resulting
in script failure due to how bash ignores `set -o errexit` when
running commands in conditionals. Adding an explicit compatibility
check ensures an obvious failure mode and path to resolution.