Add support for Maps to the reflect_codec
Support to serialize/deserialize maps with the reflect_codec. This is needed in
the subnet EMV to use it to exchange information about the upgrade configs.

This PR adds support for maps and tests it.

The internal binary format is quite simple

```
	[length: 4 bytes][key1][value1][key2][value2]
```

The length is the total number of elements and keys, therefore it should always
be a pair number.