Improve NewBatch()
NewBatch() no longer takes the height as an argument, instead it will take the
next available number from the last written.

Get() will also fail if the requested height is higher than the last height
stored. Also introduced GetLastBlock() that is a tiny wrapper on top of Get()
passing the last known height