Added `GetAllAtHeight([]byte, uint64)` and `GetKeysByPrefix()`
The first function will return all the keys that matches a given prefix and
their value as they were at the requested height. This function will return the
state of the database at a given height, through an iterator.

The second function is a auxiliary function which queries a secondary index
which is a list of all known keys. This secondary index allows the first
function do a fast lookup for all the keys that matches a given prefix. This
secondary index allows to iterate quickly about the know keys, this iterator
will do another call to load the value of each key at the requested height. Any
key that was not yet created or that was deleted will be ignored by the
iterator.