Fix key issue with same prefixes
This bug was reported by @StephenButtolph. He even provided a test suite that
passes now.

The issue happens at the `Get` function, it would just attempt to read *one*
record from the database and if the requested key is not the same it will
return not found. This bug exploited how the keys are stored, crafting a longer
key that will be stored first. When checking if the keys are the same, it would
fail with a NotFound error.

The fix was to introduce a loop in the Get function. This loop will usually
just do a single read/try *unless*:

	1. The found key is longer than the requested key
	2. *and* the found key is has the same prefix as the requested key.