Remove secondary index
Removed the secondary that was keeping track of all defined keys sorted by
prefix. Instead, an iterator is being used to get a list of all defined keys
under a prefix.

The disadvantage is discovering every key is a random read, not sequential as
before, but each should be pretty quick since each key are naturally sorted. In
contrasts inserting new data to the database is cheaper since  building the
secondary index is not longer needed.