Added a common prefix for all keys
As pointed out by @StephenButtolph, without this PR it would be possible for an
external database user to craft a key named in such a way that it would be
possible to pollute the key space with a common prefix. They won't be able to
make the database give incorrect answer, but they will be able to make us slow.

By introducing a common prefix that will be added, regardless of how the
external key is being crafted the problem is solved.