Add Batch() expected behaviour
According to @darioush[1], the Batch should work the expected way. Which means
allowing the *current* height to be updated (Write() multiple times). If the a
new block has been written, the batch writter won't be valid.

Reset() and Replay() work as expected as well. The Batch interface is
compatible with other databases implementations

[1] https://github.com/ava-labs/avalanchego/pull/1983#discussion_r1325008133