`tmpnet`: Ensure nodes are properly detached from the parent process
Previously tmpnet called `Cmd.Wait` on the node process based on a
misreading of the advice of the `Cmd.Start` to ensure that system
resources were freed. This ensured that any error in the parent
process would result in the node processes being killed.

Now, the child is properly detached by setting `Setsid` and avoiding
calling Wait. This ensures that the node restarts required when
setting up subnets (e.g. for subnet-evm or hypersdk) do not risk the
nodes of an existing network being stopped if the test suite exits
non-zero due to a test failure.