[tmpnet] Misc cleanup to support xsvm warp test PR
- update tmpnet.ReadFlagsMap to avoid returning an unnecessary pointer

- add tmpnet.InterfaceToFlagsMap helper to simplify conversion of
subnet configuration to a FlagsMap for nicer serialization

- update tmpnet.NewNodes to initialize tls and signing keys so that
nodes returned by this method can be used to extract the validator IDs
for a subnet

- switch tmpnet.Chain.Genesis back to a byte array. A previous attempt
to improve the readability of tmpnet subnet configuration by
converting to FlagsMap ignored that genesis byte format is VM-specific.

- switch tmpnet.Chain.Config back to a string. A previous attempt
to improve the readability of tmpnet subnet configuration by
converting to FlagsMap ignored that VMs need to be able to specify
their exact configuration.