[tmpnet] Add network reuse to e2e fixture
Previously, using an existing network with the e2e suite required
starting the network with the tmpnetctl cli. This has been replaced
with direct support for network reuse within the e2e fixture to better
support testing of networks with subnets and remove the need for
other repos (subnet-evm and hypersdk) to use tmpnetctl:

 - Replace the --use-existing-network flag with --reuse-network flag
 which will ensure that a suite-compatible network is started if
 not already running and reused if already running

 - Added the --stop-network flag to support stopping a network
 previously started by --reuse-network and exiting immediately without
 running tests.

tmpnetctl is left in place for now but its future is uncertain.