CI: ensure image build job is compatible with merge queue
The name of the branches used for the merge queue contain slashes, and
since docker image tags cannot include slashes, the merge queue branch
names are not valid for use as docker image tags.

To fix this, convert a branch name to an image tag by replacing any
slashes with dashes (which are allowed).