Fix tests that depend on divergent betaVirtuous/betaRogue values
Remove RecordPollDivergedVotingTest - this test requires divergent betas so that
in the binary decomposition of voting, it's possible to conditionally accept an
intermediate node and then transitively apply a vote through a block that has already
been pruned due to the accepted prefix.

Modify LastAcceptedTest - modifies the beta value to 2 and performs 2 additional polls
to test the same logic of transitively applying voting through blocks and increasing
the last accepted block by one at a time. This replaces relying on divergent betas with
recording a poll with blocks at different heights.