[e2e] Simplify pre-funded key usage
As part of refactoring the e2e tests for reuse in antithesis and load
testing, this change simplifies key usage so that tests can be
supplied a specific pre-funded key rather than retrieving their own
key at runtime. Fund usage is also minimized. It should be possible to
run most tests repeatedly for extended periods of time without running
out of keys or funds.

Previously, pre-funded keys were distributed via an http server to
ensure a given key was used by at most a single test. This was
intended to ensure compatibility with parallel test execution by
guaranteeing that no pre-funded key would ever be used concurrently.

Since ginkgo uses process-based parallization, it's sufficient to
ensure that each process is allocated a single unique key. The tests
assigned to run in a given test process are executed sequentially so a
key won't be in danger of concurrent usage.

As part of this change, the virtuous test was updated to use one
pre-funded key to fund a new set of keys. Previously 10 pre-funded
keys were required.