[antithesis] Enable workload instrumentation
The Antithesis test setups currently use regex matches on log output
to trigger fault injection (by matching on output indicating that test
setup was completed) and record desired test properties (e.g. by
matching on output indicating that an X-Chain transaction was
verified). The configuration of this matching is currently stored on
Antithesis infrastructure and only accessible to Antithesis personal.

This changeset uses the Antithesis SDK to explicitly signal events so
that the use of potentially unreliable regex matching can be avoided
and that more control over test execution and analysis can be given to
code in the avalanchego repo.