[e2e] Fix excessively verbose output from virtuous test
PR #3057 added a metrics client and switched the virtuous test to use
it. The virtuous test was previously selecting only the 2 metrics it
was interested in and printing them out on each of its 50 rounds, but
after the switch to the new client it was printout out all metrics
rather than just 2, resulting in so much log output that it was no
longer possible to view a complete e2e log in github's UI without
downloading it. This change ensures that only the 2 metrics of
interest are output.