Rename Transitive snowman to Engine snowman
In the package snow/engine/snowman, the engine is called Transitive, and resides in transitive.go.
while in the package snow/engine/avalanche it's respectively Engine and engine.go.

Moreover, in the tests of snow/engine/snowman/, the bespoken object is referred to as engine.

I therefore rename the transitive to engine to better fit the rest of the code.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>