Activate Error Driven Snowflake
This commit enables error driven snowflake and reduces the average poll time by 35% on the Fuji network.

In the classical snowflake protocol, a node issues a sequence of polls, and every poll yields a certain confidence score which is based on the number of nodes that responded and the content of the response.

If the poll contains enough responses that amplify the confidence score above a certain threshold, the poll is considered a success, and the criteria to finalize a block is collecting enough successive successful polls.

In the error driven snowflake which is described in Section 4.1 in the Frosty paper, a poll can succeed in various degrees of success: The higher the confidence score that the poll concludes, the more successful the poll is considered. In contrast to the classical snowflake protocol, the criteria for how many successful polls are required to finalize a block is now determined by the confidence score of the polls - successive polls with a higher confidence score require fewer of them to finalize, and vice versa.

Each poll consists of sending queries to a number of nodes, and collecting responses. Since nodes may be offline, slow or malicious, they might not return responses in a timely manner. In order for the polls to be efficient, there exists logic which terminates a poll early once it has reached a required level of confidence, or if enough nodes timed out such that it is evident that the required confidence level cannot be reached by waiting for further nodes.

The current snowflake code already supports the error driven variant. However, the logic that terminates the polls early currently only supports the classical snowflake with a single confidence score.
In addition, there is no way to express in the configuration the threshold for the error driven snowflake, as there is only a single confidence configuration in the configuration.

This commit introduces new configuration flags to the avalanche node which express the error driven snowflake various confidence criteria, and also changes the early termination logic to accommodate the error driven snowflake.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>