[tmpnet] Add support for checking rpcchainvm version compatibility
Previously, if the rpcchainvm version differed between avalanchego and
a VM that a node was configured to support, the node would never
become healthy and discovering this would require examining log
output. This commit adds support for optionally checking the
rpcchainvm version of a VM against the version supported by the
configured avalanchego binary.