[antithesis] Fix broken flag handling and improve image testing
The recent addition of tmpnet support for antithesis workloads broke
configuration in the actual antithesis environment due to deficiencies
in the test of the image build. This change cleans up the flag
configuration that caused the problem and updates the test of image
build to directly check the output of the docker compose project for
evidence that nodes have reported healthy rather than just waiting 30s
for a non-zero exit.