Implement traversal based early termination
This commit makes the early termination logic of the snowflake poll consider also transitive voting.
A block `b0` is transitively voted by all votes for block `b1` if `b0` is an ancestor of `b1`.

This transitive termination logic also works for shard common block prefixes.
If block `b0` is a direct ancestor of `b1` and `b2` and the latter have a shared prefix,
then the logic in this commit may not early terminate the vote if a future vote might improve the confidence
of the shared prefix, as it correspons to a snowflake instance.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>