Make snowman use snowflake directly instead of snowball
Snowball was introduced in order to improve the stability of snowflake,
by taking into account the history when returning the preference.

However, if deployed with a configuration where the preference is small enough, and a network partition
causes a 50-50 split of the stake, it can actually backfire, as described in 6e1a905.

Additionally, latest research [1] points out that snowflake suffices for snowman.

This commit removes snowball from snowman and adds a test that simulates a mixed network
which runs snowman with and without snowball, and ensures that a mixed network still converges.

[1] https://arxiv.org/abs/2404.14250

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>