Fix node initialization in test
The topological consensus tests were supposed to have the instances initialized with a random tree of blocks,
where each block points to a random set of descendants.

Instead, because of a bug that influenced the parent reference, the blocks were added to the consensus instance
as a single block with the rest of the blocks as descendants.

That's not what we want to test, so this commit fixes this and further simplifies the initialization logic.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>