Publish last accepted height through chits
This commit adds the last accepted height to the Chits message.

The intention is that by receiving the latest accepted height of each node,
a node can know whether it is straggling behind the majority of the nodes in the chain.

I ran a build with the commit on a Fuji testnet node and verified the log matches the commit.

```
[10-02|19:59:48.312] VERBO <P Chain> snowman/engine.go:362 called Chits for the block {"nodeID": "NodeID-1AFBjE7UUM1AWWA1HtMGreRrvKwTh9Su", "requestID": 335, "preferredID": "zmoZ2RPy2mkx1ozSbhen9Ggh9boTewBCa6bgc5FnQu7d2umrf", "preferredIDAtHeight": "zmoZ2RPy2mkx1ozSbhen9Ggh9boTewBCa6bgc5FnQu7d2umrf", "acceptedID": "zmoZ2RPy2mkx1ozSbhen9Ggh9boTewBCa6bgc5FnQu7d2umrf", "acceptedHeight": 0}
```

Also verified on its grafana dashboard that it participates in consensus and that there are no failing health checks.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>