[testing] Enable bootstrap testing of partial sync
Previously there were only 2 modes for bootstrap testing - state sync
and full sync. This change enables partial sync which only processes
the history of the P-Chain and is expected to be the default mode for
L1 validators.

Rather than adding a new field to the serialized test configuration,
the existing `StateSyncEnabled` field becomes `SyncMode` with 3
values: `full`, `state` and `partial`.