[tmpnet] Set an explicit `instance` label for logs and metrics
Previously, the configuuration for promtail and prometheus did not set
an explicit `instance` label. This resulted in the default of `ip:port`
being used, and since that could change on restart, it was necessary
to try to ensure a stable port. Setting an explicit and unique value
for `instance` avoids needing to ensure a stable port.