[docker] Enable image builds from a git worktree
This change enables docker image builds from [git
worktrees](https://git-scm.com/docs/git-worktree). This required
discoverying the git commit used to configure the avalanchego build
script in advance of image build and providing it as an argument to
image build.

Previously, it was impossible to build docker images from a git
worktree:

 - the avalanchego build script needed to discover the git commit to
 configure the binary with for traceability

 - the build script assumed a .git directory to discover the git
 commit

 - the dockerfile copied the working tree into the image prior to
 running the avalanchego build script

 - if the working tree was for a git worktree, the resulting copy
 wouldn't have a .git directory to discovery the git commit from

  - instead of a regular .git directory, a non-primary worktree has a
  .git file referencing a path in the primary worktree's .git
  directory.