Fail fast in tests if avalancheGo executable isn't an absolute path
Currently when running ginkgo tests with --avalanchego flag with a relative path,
the tests fail when trying to ascertain the rpcvm plugin path, because they execute the binary
without an absolute path, and the binary path is evaluated from a wrong location.

In order to make the underlying issue clear, I added a check that fails the test
and warns that an absolute path must be used.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>