[ci] Fix metrics link annotation emitted for e2e and upgrade jobs
Previously, the upgrade job was configured to filter the metrics link
annotation with `network_owner=avalanchego-e2e`. Not sure how it ended
up that way, because only the e2e job should be so filtered. This was
preventing the metrics link for both jobs from displaying results
properly. This didn't effect links emitted as part of network start,
only the link emitted as an annotation.

Also updated tmpnet README to better document the dashboard link annotation.