[docker] Update all images to debian12/bookworm
With the EOL of debian11-based Ubuntu 20.04/focal and the deprecation
and impending removal of 20.04 runners from github actions, an upgrade
to our image userland to debian12/bookworm is ensure our images use a
userland that is being tested with the debian12-baesd 22.04.

This will also ensure compatibility of subnet-evm binaries built with
goreleaser on 22.04 runners (and so requiring a debian12 userland)
with avalanchego and subnet-evm images that will now also have a
debian12 userland.