Fix flake TestIngressConnCount
The test spawns three instances of test network one after the other.
In an overwhelming number of cases, they connect to each other according to their order of creation.

However, in some rare cases, they connect to each other out of order.

I made the test use set semantics instead of list semantics in order to not rely on their order
of creation.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>