[ci] Enable run-monitored-tmpnet-cmd to use a remote flake file
Previously the run-monitored-tmpnet-cmd custom action assumed a local
flake file. This required any user of the action to define a local
flake file even if they had no local need for one. Now the command
will attempt to use a local flake file but if one is not present, fall
back to invoking the avalanchego flake file at the vendored version.