[tmpnet] Switch FlagsMap from map[string]any to map[string]string
`map[string]any` complicates comparison of values converted to
strings (e.g. when node configuration is provided as env vars to a
kube pod) to their original golang form. Switching `FlagsMap` to use
`map[string]string` forces the decision of how to serialize on the
developer to ensure that all subsequent usage will be consistent.

This change also updates chain and monitor configuration to stop using
FlagsMap in favor of locally-appropriate maps.