[tmpnet] Replace TestEnvironment.URIs with GetLocalNodeURIs()
Previously it was possible to use TestEnvironment.URIs to access node
API endpoints. If a node is running in a kube cluster, however, it
won't be accessible from outside that cluster. Adding
TestEnvironment.GetLocalNodeURIs and updating callers to use that
ensures URIs that will work regardless of the location of nodes and
test workloads.