[tmpnet] Enure node config is saved on restart for both runtimes
The addition of the kube node runtime added a `Restart` method to the
`NodeRuntime` interface to support runtime-specific restart
behavior. This introduced a disparity in restart behavior between the
process and kube runtimes. The kube runtime was saving the node
configuration before restarting, but the process runtime was not.

When starting a network configured with one or more subnets using the
process runtime, this resulted in `track-subnets` for the bootstrap
node not being serialized to disk. Only the bootstrap node was
affected because only it was being restarted after having
`track-subnets` set. The remaining nodes were started with
`Network.StartNode` - which ensures serialization - after
`track-subnets` was set.

This broke the xsvm and subnet-evm antithesis testing because both
depend on the process runtime to initialize a target network and the
bootstrap node would end up without a `track-subnets` value. This
resulted in 404 responses when trying to query a subnet's rest
endpoint.

The fix is to ensure that `Node.Restart` serializes node configuration
regardless of runtime.