[tmpnet] Enable externally accessible URIs for kube-hosted nodes
Previously, access from outside of a kube cluster was enabled by
port-forwarding through the kube API. This approach turned out
incompatible with load testing because it greatly limited the rate at
which transactions could be sent.

The port-forwarding is replaced with nginx+ingress to ensure minimum
overhead when running outside of a kube cluster.