perf: remove some unecessary allocs during serialization (#965)
In #831, it was mentioned that the benchmark for serialization showed we
were slower than serde for a full node. I noticed when walking though
the code that we were allocating a fixed sized array in a tight loop.
This alloc was also unecessary as the `TrieHash` already implemented
`AsRef<[u8]>`.

Removing this alloc and copying directly from the existing hash
significantly improved the benchmarks. There is a small and significant
regression in the leaf benchmark; however, in aggregate the performance
overall has improved.

https://gist.github.com/demosdemon/fa1258e96c8b09851b18e84c45005a28