[tmpnet] Run kube load test under a service account to validate RBAC
Running tmpnet in a pod (as is intended for the load test) requires
that the pod run under a service account with sufficient permissions
for the tmpnet kube runtime to operate. To validate the RBAC
configuration, the task running the load test under kube has been
updated to execute with the set of RBAC permissions that tmpnet requires.