feat(delayed-persist): Part 1: Roots may be in mem (#1041)
Introduces a new type, `MaybePersistedNode`, which is a node which is
either a SharedNode or a LinearAddress. This new type will allow writes
to occur while readers are still reading this revision using an
`ArcSwap`.

Note that this does not actually put any additional nodes in memory, it
just sets things up so they could be in memory.

Additional changes needed for this feature to be complete are:

- [ ] BranchNode children can now be Node, LinearAddress or a
MaybePersistedNode
- [ ] When converting a `MutableProposal` to an `ImmutableProposal`,
don't actually allocate space, just create `SharedNode`s from the
`Node`s.
- [ ] Remove `NodeStoreHeader` from `NodeStore`. The header should move
into the `RevisionManager`.
- [ ] Allocate new nodes from the recently-freed nodes from an expired
revision, then from the freelist. This can be done by maintaining a
local freelist from the nodes deleted in the expiring revision and
falling back to the actual freelist. Deleted nodes that are not reused
must still be pushed onto the freelist.