Fix flake in TestNotifierReSubscribeAtPrefChange
The subscriber notifies the test to cancel the context via closing the signal channel,
which in turn makes the test to call CheckForEvent on the notification forwarder.

However, the test goroutine proceeds to send a message down to the messages channel right after.
The subscription function, however, doesn't prioritize exiting and may read the message which was intended
for the next iteration of the forwarder.

This commit simply forces the subscription function to return early in its first iteration and not proceed
to read the message.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>