feat(checker): check for misaligned stored areas (#1046)
This PR adds checks for stored area misalignment. We only focus on areas
that are not leaked (areas tracked in the trie or in the free list). As
long as the areas not leaked are aligned, leaked areas will be aligned.

When a misaligned area is identified, we return not only the address of
this area but also who points to that area directly. If this area is
allocated for a trie node, the root or the parent trie node will be
returned. If this area is free, the free list head or the parent of this
free area is returned. We return the parent because, if we were to fix
the misalignment issue (by migrating to another area or remove it from
the free list), the parent pointer needs to be updated accordingly.

As a side effect, `FreeListIterator` will return the parent of the free
list item. This design can help checker fix various issues it sees when
traversing the free list but will add some overhead if it is used for
other purposes.