feat(deferred-allocate): UnpersistedIterator (#1060)
An UnpersistedIterator iterates through a NodeStore looking for
unpersisted nodes. Unpersisted nodes are only of the type
MaybeUnpersisted that are not yet persisted.

Note that it's impossible to have a persisted parent for a
MaybeUnpersistedNode, so it's relatively easy to find the unpersisted
ones.

There are tests that verify that the nodes are returned in the order
that they will need persisting (that is, depth first).

---------

Signed-off-by: Ron Kuris <swcafe@gmail.com>
Co-authored-by: Suyan Qu <36519575+qusuyan@users.noreply.github.com>