Throttle platformVM block building until normal ops
When the platformVM is bootstrapping, WaitForEvent() is called on it by the notification forwarder,
which forwards the events to the handler, which in turn routes them to the appropriate engine according to the state.

The platformVM cannot know when it should propose a block until it has finished bootstrapping, and as a result it will fail in attempting.

If the proposerVM selects the platformVM for block building before it has finished bootstrapping, it will result in the notifier in a spinning loop
until it has finished bootstrapping.

This fix makes the platformVM throttle its WaitForEvent() call until it has finished bootstrapping.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>