docs: update ffi/README.md to include configs, metrics, and logs (#1111)
After many changes to the FFI pattern, we should maintain an up-to-date
guide on what the different configurations mean, and how to use them.
Additionally, the docs are now available in the crate's documentation.
To check this locally, run the following commands from the root of the
repo:

```bash
cd ffi
cargo doc --open
```
Closes #1078