feat: add `Children<T>` type alias (#1123)
This commit introduces a new type alias `Children<T>` which is a
fixed-size array of `Option<T>`. This is convenient for all of the
places where we represent the children of a branch node in the trie.

Also included is a method to construct the array in const context,
allowing for easy creation of `Children<T>` instances.

Perhaps most controversial in this change will be the changes to
`children_addresses` and `children_hashes`. Instead of returning an
iterator, they instead return the array of requested values. The
implementation of this is also of particular note because I `zip` over
an `iter_mut` preventing the need for any array indexing like was
previously used to construct these arrays.

This change was primarily motivated by my range proof verification work.
I was touching code that indicated differences with branch_factor_256,
so I expanded my clippy checks to ensure that the code compiles with the
feature enabled and discovered that it didn't unrelated to my changes.
The `Default::default()` calls work for the children array when
`branch_factor_256` is disabled because rust stdlib has an
[implementation for arrays of up to 32
elements](https://doc.rust-lang.org/std/primitive.array.html#:~:text=Arrays%20of%20sizes,to%20size%2032.).
This is due to a limitation of the type system that prevents a generic
implementation over the generic size.