chore(release): prepare for v0.0.11 (#1216)
This PR bumps the crate versions to `0.0.11` and updates the CHANGELOG
generation to also include links to the pull request. We could use the
`github` integration in git-cliff to ensure we get the actual pull
request link but that adds a significant amount of time to the
generation time because of all of the github API requests. It also
requires the user have `gh` installed and be authenticated.