chore: only allocate the area needed (#1217)
As shown from statistics, Firewood currently has a lot of internal
fragmentations because it will allocate any large enough free area from
the free list, often times in which the allocated area is several times
bigger than the area needed. This PR mitigates this issue by only 
allocate the area needed to store the node but not larger areas. It 
reduces storage overhead to 1/4.