feat(async-removal): Phase 4 - Make `DbView` synchronous (#1219)
This also removes all final `async` and `stream` usages in the code,
with the exception of `firewood-benchmark` which must use tokio for
handling the prometheus metrics listener.

This finalizes the `async` removal process.

# Why `ArcDynDbView`?

You may wonder why I replaced `Box<dyn DbViewSyncBytes>` with `Arc<dyn
DybDbView ...>`. This is because in both cases where it was used, we
already had an Arc object that the compiler could coerce into our
`Arc<dyn ...>` variant without adding an additional box into the mix.
This removes one layer of pointer indirection.

Fixes: #798