feat(ffi-refactor): add OwnedSlice and OwnedBytes (2/8) (#1223)
This is similar to the BorrowedBytes from #1174 but for bytes returns by
Rust back to Go.

This does not actually use the new type yet as there are too many places
that need to be updated. However, this introduces the new type and it's
free method so that we can start using it in the future.

Depends On: #1222