feat(ffi-refactor): refactor how hash values are returned (5/8) (#1226)
This way we avoid allocating just to return the hash value.

A 32-byte hash value is the same size as 4 words and the allocated
buffer would be returning 2 words. If we skip the heap allocation, we
can return all 4 words on the stack directly and avoid the heap
allocation (on the rust side). The Go side later allocates for the
slice, but this could be removed later if the interface was build around
a fixed-size 32 byte array.

Depends On: #1225