test(firewood): use ctor section to init logger for all tests (#1277)
This uses the `ctor` library to install a function in the binary `ctor`
section that initializes the logger for all tests.

This makes it so we don't need to manually pepper in logger initializers
for all tests.

This isn't really safe for production level code and should only be
enabled for testing.