feat(ffi-refactor): replace sequence id with pointer to proposals (8/8) (#1221)
This is the final PR in the series to refactor the FFI layer to be more
rust and also a bit more Gooey.

This replaces the serial atomic id for proposals with a pointer to the
actual structures. The wrapper that includes the db handle is only
necessary in order to do the cached view shenanigans. I will come back
to that later.