chore(release): prepare for v0.0.13 (#1317)
The changelog is a little messed up because when I tagged `v0.0.12`, I
accidentally did it from the branch and not the commit that was squashed
and merged into main. When git-cliff walks the history, it fails to find
the tag thus 0.0.12 technically didn't happen. Correcting this in order
to correct the changelog would require re-tagging and that would be very
disruptive. It's unfortunately better to pretend 0.0.12 didn't happen.
This is also why I amended the release doc.