[tools] Add depctl tool to simplify consumption of local repos
In the absence of a monorepo, individual
repos (avalanchego,coreth,subnet-evm,firewood) all need a way of
testing pre-merge with their dependent repos. This is currently
accomplished in coreth and subnet-evm via bash scripts that simplify
maintaining the version of their avalanchego dependency and cloning
avalanchego locally to perform testing.

depctl is a new tool written in golang to simplify maintenance of this
shared functionality. Since all these repos depend on golang, such a
tool can be invoked with `go run`. depctl is more flexible than the
previous scripts to better support firewood testing.