refactor(db): `TestDb` constructor (#1351)
I recently learned from @demosdemon that in Rust, the default
constructor of a type is generally called `new()` while additional
constructors are called `with_X()`.

This PR refactors the `TestDb` code to have a `new()` constructor with
this function being called through the database tests instead of
`testdb()`. This helps break up #1346 as we'll just then need to add a
`with_mockstore()` method to `TestDb` to allow for injecting instances
of `MockStore` to `TestDb`.