feat(ffi): implement revision handles and expose to FFI layer (#1326)
This PR introduces revision handles that hold a revision across the FFI
while you use it, so it won’t be removed mid-operation. This should
improve concurrency and avoids errors as mentioned in #1008 . It also
makes iterator's design simpler, we can create them from a held
RevisionHandle with no tricky lifetimes.

This updates the existing Go bindings so `Database.Revision` returns a
Revision with `Get` and `Drop` methods. Necessary changes to tests are
applied.
  
Closes #1008.