feat(1/7): add U4 type to be used as a path component (#1336)
This change introduces a `U4` type that uses an underlying enum with
only 16 states in order to enable niche optimizations. This change does
not yet use the type within code as that will cause a large amount of
changes that will make reviewing this difficult. The focus of this
change is the `U4` type itself and nothing outside of that. This type
also is not a `PathComponent`, which will appear in the next Pull
Request.

## Memory Layout

The niche optimizations apply to memory layout allowing for `Option<U4>`
to have the same size and alignment of `U4`, meaning there is no
additional storage to keep the Option discriminant.

## Array Bounds Checking

The niche optimizations also apply to bounds checking when indexing into
a fixed size array, like we do for the `Children<_>` type alias. See
<https://rust.godbolt.org/z/cqdz1n9d5>

This change contributes to the resolution of #1230.