feat(4/7): add SplitPath trait (#1339)
One of the most common operations on paths is finding the longest common
prefix between two paths. This is useful for a variety of reasons, such
as determining if one path is a subpath of another, or finding the
relative path between two paths.

A normal `TriePath` implementation does not provide a way to do this
because that trait should remain `?Sized`. Whereas, splitting the path
requires a sized type in order to produce a new instance of the path
when splitting.

This change adds a `SplitPath` trait that provides the ability to split
a path. Also added is `IntoSplitPath` which is a conversion trait to
yield a splittable path. This also acts as the borrower for owned paths
to yield shared references to the path.