feat: Use PathComponent in proofs (#1359)
This change primarily replaces the soon to be removed `Path` struct with
`PathBuf` (which is a type alias for the smallvec of `PathComponent`)
within the proof structure and related hashing traits. Changing only the
proof and related components was less invasive changing `LeafNode`,
`BranchNode`, or the iterator states. Those changes will come in future
updates.