feat: replace NodeAndPrefix with HashableShunt (#1362)
The `NodeAndPrefix` struct was too limiting in that it required a full
node from the storage layer. In order to verify proofs, we only need the
individual components of the hash preimage and not the full node.

This change introduces `HashableShunt` which is a more flexible
abstraction that can be constructed from either a full node or the
individual components. This allows us to create shunts for hashing
directly from proof data without needing to reconstruct full nodes.