feat: Parallel updates to Merkle trie (v2) (#1258)
Second implementation of adding parallel updates to a Merkle trie. In
this approach, each child of the root node is a separate trie that can
be modified independently by a separate worker thread. Four steps are
needed to create a parallel proposal:

1. Prepare: The trie is modified such that the root is always a branch
with an empty partial path. This prevents any update from replacing the
root node with one of its children, which is necessary to allow the
sub-tries to be operated on independently.
2. Split: The updates from a batch are split and sent to different
workers based on the first nibble of their corresponding keys.
3. Merge: After completing the batch, each worker returns the root of
its sub-trie to the main thread. The main thread attaches the sub-trie
roots to the children array of the root node.
4. Post-process: The Merkle trie is returned to its canonical form. This
eliminates cases where a trie has a root node with no value and only one
single child, or a trie has a root branch node with no value and no
children.

The main changes are in parallel.rs. A test case (test_propose_parallel)
has been added to test parallel insertion/deletion.

---------

Signed-off-by: bernard-avalabs <53795885+bernard-avalabs@users.noreply.github.com>