fix: Revert "fix: Revert "feat: Parallel updates to Merkle trie (v2)"" (#1374)
Reverts ava-labs/firewood#1372

At around 8:30 pm Eastern on Oct 16, 2025 my merged to main PR was
failing `firewood-merkle-differential-fuzz`. My PR passed this test
prior to merging but it failed consistently with an out of memory error
after merging. A revert PR was merged to see if that would fix this
issue, but the test continued to fail after merging the revert PR.

Around an hour after merging the revert PR, the
`firewood-merkle-differential-fuzz` began passing consistently. I've
also reran the test at both head and before the revert (commit hash
49ab032c4112f65637f0d8f9dadbe98cd1bca6e7) this morning (Oct 17) and it
is passing consistently.

I suspect it was a CI runner/infrastructure issue that caused the test
to fail. This PR is to revert the revert PR from last night.