chore(db/manager): add `MockStore` (#1346)
As part of #1299, this PR introduces the `RootStore` trait and makes
both the revision manager and database generic over `RootStore`. The
revision manager, in particular, is modified to make calls to its
`RootStore` as described in the `RootStore` design doc. It's expected
that the archival state datastore will implement this `RootStore` trait
in production.

This PR also introduces two implementations of `RootStore`:
- `NoOpStore`: a no-op implementation of `RootStore` that is used in all
contexts outside of testing
- `MockStore`: a mock implementation of `RootStore` that is used for
testing the `RootStore` API.

This PR also introduces two database tests to test the `RootStore` API.

---------

Co-authored-by: Brandon LeBlanc <brandon.leblanc@avalabs.org>