feat: Parallel hashing of Merkle trie (#1303)
This PR builds on https://github.com/ava-labs/firewood/pull/1258 to
support parallel hashing. The workers performing parallel inserts will
now perform hashing on their subtrie when they receive the Done message
from the main thread. This is safe even if the trie is modified after
post-processing because there is only one case in which any of the
subtries is modified during post-processing (the root only has one child
and no value). For this case, the updated root will be rehashed.

The function hash_help has modified to take a path to allow the workers
to correctly hash their subtrie by accounting for their child index.

---------

Signed-off-by: bernard-avalabs <53795885+bernard-avalabs@users.noreply.github.com>