chore(ffi/firewood): remove `RootStore` generics (#1388)
While going through a first pass of adding `ProductionStore`, I found
that adding a `RootStore` generic to `Db` resulted in the generic
bubbling up to the `ffi` layer, which caused a myriad of issues from the
Golang side (and a massive refactoring).

This PR removes the `RootStore` generic and instead sets the
`root_store` field of the `RevisionManager` to be a `Box` pointer.