ci: Update ffi build check to configure cargo with same MAKEFLAGS as nix (#1392)
As per a [recent
flake](https://github.com/ava-labs/firewood/actions/runs/18783578425/job/53595874673?pr=1388)
it's not enough to configure the nix build with `MAKEFLAGS='-j1'` to
avoid jmalloc non-determinism. The cargo build invoked in the nix shell
also needs to be configured this way or CI runs on x86 runners run the
risk of non-deterministic builds.